libdfp: New framework for tests
This patch adds a new framework to create tests for libdfp. Instead of
static code C tests the new framework uses descriptive files with input
and expected output. It will be processed by a script that will actually
generate code.

A example test description is as follow:

  # name quantum
  # arg1 decimal
  # ret  decimal

  Inf                    Inf
  -Inf                   Inf
  NaN                    NaN
  sNaN                   sNaN

  DEC_MAX                1E+DEC_MAX_EXP
  DEC_MIN                1E-DEC_MIN_EXP
  DEC_SUBNORMAL_MIN      1E-DEC_MIN_EXP_SUB

  -DEC_MAX               1E+DEC_MAX_EXP
  -DEC_MIN               1E-DEC_MIN_EXP
  -DEC_SUBNORMAL_MIN     1E-DEC_MIN_EXP_SUB

Initial lines describe the function (in the case quantum that has one
_Decimal as argument and return a _Decimal). Each line following
descriptions will be translated to a function call and its return will
be checked against expected value.

The 'decimal' directive will use a _Decimal type according to script
input argument. This makes the test description file independent of the
decimal type. Some directives will be set in accord of decimal type,
like DEC_MAX to DEC32_MAX for _Decimal32.

Some tests should be run only for certain decimal types. To do so, the
directive decimalN can be used as:

  20.971525              0.000001               decimal64 decimal128

The test will run only for _Decimal64 and _Decimal128.