libdfp: base-math symbol hidden_def
This patch makes all the libdfp _Decimal operatations (like sum,
multiplication, cast, etc.) have a hidden definition.  This is be able
to create symbol alias and avoid intra-library PLT calls for compiler
generated calls for _Decimal operations.

For instance, on the code:

_Decimal32 foo (_Decimal32 a, _Decimal b)
{
  return a + b;
}

The compiler may emit a call to '__<bid/dpd>_addsd3' if the plaform does
not support hardware instructions.  Since libdfp provides its own
implementation for the symbol, if we dot an alias in the form:

asm ("__<bid/dpd>_addsd3 = <internal symbol name>")

The compiler/linker will eventually create a PLT calls to the add
implementation instead of a local one.