libdfp: Correct expected output for quantize tests
This patch fixes the expected output for some tests for _Decimal32 where
the coefficient after rounding/adjustment should overflow and thus the
result should be a NaN value.

Current powerpc code does not handle these cases because it first
convert the _Decimal32 value to _Decimal64 (since the platform does not
have native support for _Decimal32).  The patch also adds a note on
powerpc code to fix these cases.