Adds printf tests with two % modifiers
Issue #1 states that using both %f and %Hf in the same printf call results in
error.

According to the description of the issue, %f was used with a long double
argument, instead of with %Lf, as is required by printf (see 'man 3 printf').

Trying to print a long double with %f does cause the error reported. However,
if %Lf is used, the outputs match the expected.

This commit copies the tests in tests/test-printf.c, however it modifies the
parameters to include an extra %Lf modifier.