Adds wrong test case for issue #1
This commit adds a wrong test case that addresses issue #1. The test case is
wrong, because it passes the %f modifier to printf along with a long double
argument.

The printf man page (man 3 printf) states that long double arguments should be
printed with %L followed by a, A, e, E, f, F, g, or G, e.g.: %Lf.