Fix edge case with strtod*
If requested to parse a number of the normalized form x * 10 ^ EMAX,
a valid looking tiny number may be returned without setting errno.

Some such numbers are 4e97, 4e385, or 4e6145.

The quick check for an obvious overflow allowed these numbers to be
parsed. With soft-dfp, setexp*() will now set the c field to infinity
if a slightly large exponent is passed into it. Note, this will still
crash/return a bogus value for certain large exponents which should
not be encountered.

For hardware implementations, errno will now be set if such a number
is encountered. In addition, a few extra test cases are added to
check 32 and 128b versions.

Signed-off-by: Paul E. Murphy <murphyp@linux.vnet.ibm.com>