Add most ti <-> Decimal conversions
Credit goes to Steven Munroe <munroesj@linux.vnet.ibm.com>
for the float{uns,}titd routines.  ti -> {dd,sd} is still
a work in progress.

These still leaves plenty of room for optimization, but these
are a functional first pass.

This does a bit of build system shuffling as TImode is not
supported on all platforms, and thus a separate ABI list
needs to be kept for these cases.

Signed-off-by: Paul E. Murphy <murphyp@linux.vnet.ibm.com>
Signed-off-by: Steven Munroe <munroesj@linux.vnet.ibm.com>