Synchronize plt checking scripts
Debian exposed a bug whereby certain variants of readelf will
append the symbol version. This was fixed in glibc some time
ago. These are synchronized with glibc commit d3d9c95.

One minor change is made, the regex on line 76 of localplt.awk
is left unchanged as the build directory structure is a bit
different than glibc.

This fixes Issue #38.

Signed-off-by: Paul E. Murphy <murphyp@linux.vnet.ibm.com>