s390: Fix left_justify
This was very broken. After digging up the z/Architecture
Principles of Operation, I was able to determine what was
going on.

z support is very similar to POWER, but it has a lot more
flexibility in how things are done.

Note ???x??? instructions operate on _Decimal128, and
     ???d??? operate on _Decimal64/32. Similar to POWER,
_Decimal32 is emulated through _Decimal64 instructions.

Finally, the "Round toward zero" is the constant 5 or
9, not 1 like POWER.

Signed-off-by: Paul E. Murphy <murphyp@linux.vnet.ibm.com>