libdfp: Use correct variant d32|64|128 of __isfinite instead of always __isfinited32 in nearbyint.
The testcase test-nearbyint-d128 was failing due to segmentation fault on s390.
Because of always using __isfinited32, the _Decimal128 was casted to _Decimal32
by calling trunctdsd2(), which leads to endless recursive call to itself.

The testcase test-nearbyintd128 is now passing.
Casting from _Decimal128 to _Decimal32 is still producing a segfault,
which will be fixed in another commit.

Signed-off-by: Stefan Liebler <stli@linux.vnet.ibm.com>