S390: Add support for newer s390 cpus.
This patch adds support for configure option --with-cpu=z196|zEC12|z13.
If --with-cpu is not specified, it is checked, if the build-machine supports
hardware dfp and activates it. If not, soft-dfp is used.
The submachine level is not set explicitely in this case, so that the
compiler uses his minimum -march=... level.
Note: On s390 are no submachine folders in sysdeps/s390/.

Signed-off-by: Stefan Liebler <stli@linux.vnet.ibm.com>