libdecnumber: Fix broken magic constant
A bogus bit was being set in the exponent leading to strange,
incorrect exponents.

Note, BID encoded binary mantissa's always contain an
implicit prefix.  For small values, it is always 0b0, for
large values it is is 0b100.

To signify the 0b100 prefix, the value 0b11 is appended
after the sign bit, and the exponent is shifted two bits.

This was noted in, and resolves Issue #49.

Signed-off-by: Paul E. Murphy <murphyp@linux.vnet.ibm.com>