Fix subnormals for decimal to binary conversions.
This commit fixes the incorrect result when use "to nearest" round mode in a
underflow result generated by a decimal to binary float/double cast. It should
return -0.0/0.0 if the absolute value of the result is less than or equal to
half the least subnormal or subnormal otherwise (task #32).

This commit also fixes a wrongly produced zero instead a subnormal. The
underflow case should be checked by the minimum subnormal exponent
(-324 and -45) respectively instead the minumum normal exponent (task #30).

Signed-off-by: Rogerio Alves <rcardoso@linux.vnet.ibm.com>