S390: Add configure check in order to determine if gcc supports.
If libdfp is configured on s390 without specifying --with-cpu,
then /proc/cpuinfo was used to determine if dfp support is available or not.
But if the used compiler defaults to a cpu before z9-ec,
then the build will fail.

Now a new configure check is testing if the dfp-instruction "pfpo"
(introduced with z9-ec) is used by the compiler for converting
dfp to bfp values.

This fixes issue #62.

Signed-off-by: Stefan Liebler <stli@linux.vnet.ibm.com>