Add more under/overflow tests for dfp to bfp conversion.
It seems as the tests use incorrect input values. See more information in issue #71.
According to the comment from Joseph Myers in "Conversions incorrect on underflow #32",
"This fix looks wrong for FE_TONEAREST mode, when conversion of a value with absolute
value less than or equal to half the least subnormal in the target type should
produce +/- 0, not +/- the least subnormal."

The following min-values are defined for the bfp types on s390x:
define __FLT_DENORM_MIN__ 1.40129846432481707092372958328991613e-45F
define __DBL_DENORM_MIN__ ((double)4.94065645841246544176568792868221372e-324L)
define __LDBL_DENORM_MIN__ 6.47517511943802511092443895822764655e-4966L

=> half of the least subnormal (__DBL_DENORM_MIN__) = 2.4703282292062327e-324.
And not 2.4...26e-325.
=> half of the least subnormal (__FLT_DENORM_MIN__) = 7.0064923e-46
And not 7.0...22e-46.

This commit adds some more values for under / overflow tests.
and it adds tests for the remaining conversions:
-dfp128 -> bfp128
-dfp64 -> bfp64
-dfp64 -> bfp32
-dfp32 -> bfp32

Note: On s390x (with hardware-dfp), all those tests are passing if the
pfpo instruction is rounding according the BFP-rounding-mode.
Currently GCC is emitting the pfpo instruction with rounding according
the DFP-rounding-mode.

Signed-off-by: Stefan Liebler <stli@linux.vnet.ibm.com>