Add inexact-exception tests for rint/nearbyint tests.
This patch adds support for exception testing. Therefore a single testcase
in test-input-file has to be extended by an exception-arg. Currently you can
only specify INEXACT_EXCEPTION or NO_INEXACT_EXCEPTION. Further exceptions
have to be implemented if needed.

This patch adjusts the tests for nearbyint/rint in order to check if the
inexact-exception occured or not.
The implementation of rint was adjusted in order to raise this exception
if the value has changed by rounding.

While adding exception-support in gen-libdfp-test.py an issue in the
decimal type list was found. In case of specifying "decimal64 decimal128",
the test was only executed for the first one. Fixing this issue leads
to failures in test-quantize and test-quantum.

Signed-off-by: Stefan Liebler <stli@linux.ibm.com>