Enhance fpclassify tests.
Some tests are enhanced in order to test all classes with
positive and negative values.

Furthermore, there was an issue in powerpc implementation
in fpclassifyd32 as the input value is promoted to _Decimal64
in order to use the _Decimal64 test-data-class instruction.
The check if the value is within the range of _Decimal32
subnormal values was faulty and is now fixed.

Signed-off-by: Stefan Liebler <stli@linux.ibm.com>