Add isnormal tests.
This patch adds isnormal-testcases.

Note: The return value is specified as bool as the powerpc
implementation for isnormald128 returns 1 or -1 depending on the sign!

But there was also an issue in powerpc implementation of isnormald32.
The _Decimal32 value is promoted to _Decimal64 in order to use the
_Decimal64 test-data-class instruction to check wether the value is normal.
Thus there needs to be an extra check if the value is within the range of
_Decimal32 subnormals. (Compare to powerpc fpclassifyd32).

Signed-off-by: Stefan Liebler <stli@linux.ibm.com>