[RFC] Partially fix some of ov(und)erflow issues.
This commit is just intend to start a discussion about the underflow/overflow
issues on libdfp. As pointed out my last commit was returning some incorrect
results. The problem is that I don't think we should do a switch case and
test if the conversion results on a underflow/overflow. The division should
at end of code should round convert and return the correct result. Althought
we may return early if it's obvious overflow/underflor - if the exponent is
bigger than the max exponent (overflow) or less than min subnormal exponent
(underflow).

Those functions basically truncates the number and returns the exponent and a
SRC_TYPE (Decimal) after that it converts this to a DI integer to get the
mantissa as a long long int.

The function who truncates the number also sums 1 to the exponent. So,
for instance if the number is converted to a double an the exponent is
308 (MAX_DBL_10_EXP) the function will return 309.

My questions is if we should test if exp is bigger than 308 or if it
is bigger than 309 (MAX_DBL_10_EXP + 1). We test it after or before
the number was truncated?

This commit partially fix some errors on the new tests but, there's still
failures.