powerpc: Ignore tests that are known to fail on soft-dfp and POWER hard-dfp
The test tests/test-cast-to-underflow has 36 known failures.
The fix for these failures are being tracked at PRs #32, #71 and #84.
Having these failures makes it hard for newcomers, downstream and
continuous integration systems to detect when a failure is already known
or not.
This patch ignores the 36 tests known to fail on soft-dfp and POWER
hard-dfp.  It affects only ppc, ppc64 and ppc64le.

Signed-off-by: Tulio Magno Quites Machado Filho <tuliom@linux.ibm.com>