Show details about test fails in output of make check.
Currently there is no indication which tests fail in an
invocation of "make check".

This patch creates one <test-name>.test-result file per test
with an indication like PASS or FAIL.  Furthermore the file tests.sum
is created with one line per test in order to indicate the test result.

At the end of "make check" the failing tests are listed and the
corresponding out files are dumped.

This way you'll get an indication of failing tests even if you can
just view the log file instead of having access to the build-directory.

The scripts and some parts of the Makefile comes from glibc in order to
be more equal.

This change also allows to mark an test as:
-XFAIL or XPASS: Add a Makefile variable like test-xfail-<testname> = yes.
-UNSUPPORTED: The testcase has to return with exit-code 77.

Signed-off-by: Stefan Liebler <stli@linux.ibm.com>