Integrate libdecnumber into build
libdecnumber is very tightly coupled to libdfp. It implements most
of the transcedental functions. The choice of libbid or libdecnumber
is not really a choice. The former implements soft-dfp routines for
the GCC generated stubs today. It relies on more accurate radix
conversions to leverage BFP transcendental functions.

This will hopefully ease future maintenance as the recursive make
only complicates an otherwise simple library.

likewise, take the copy of gstdint.h used by gcc. It is a trivial
redirect to stdint.h.

Signed-off-by: Paul E. Murphy <murp@ibm.com>