Creating new nextup and nextdown functions
Both functions are just the nextafter towards +- Infinity.
But nextafter was wrong, before there was an epsilon constant and the
number was incremented/decremented by it.
The right way to implement is to inc/dec the smallest unit possible for
input number.
It was necessary to fix the left_justify function because it was just
ignoring subnormal values and to get the smallest unit possible of a
number you need to "left justify" it.
I had to replicate this same behaviour for soft-dfp.

Signed-off-by: Pedro Caldeira <pedro.caldeira@ibm.com>
Signed-off-by: Tulio Magno Quites Machado Filho <tuliom@linux.ibm.com>