sin/cos: more correctly compute the requested value
decNumber is not really built to dynamically switch precisions
during a user-facing calculation.

It generally handles it gracefully except when storing to a
reduced precision operand.

This is overcome by stashing key context values, and bumping to
the 60 digits of precision otherwise implied by these functions,
otherwise, the operations are carried out using the input context
which is usually far less. This can result in up to 7 figures of
rounding error with decimal32 arguments (e.g around pi/2).

Notably, the linchpin is assumption that reseting the precision and
performing a "no-op" addition (e.g +0) will correctly truncate into
the final destination operand.

Likewise, I claim this is more correct, but this hinges on the
assumption that 60 digits is sufficient precision to get to <= 1ULP.

Likewise, simplify the factorial arithmetic a little. With expanded
exponent range, we can compute the factorial well beyond 63 terms,
and the product of the highest two values can always be computed
via int32, so maybe we save a little bit there.

Signed-off-by: Paul E. Murphy <murp@ibm.com>