Remove unused or fixed configury
OPTION_EGLIBC_LOCALE_CODE is always defined. Make it so, and remove
the long unused option.

Likewise, remove some degenerate autoconf for confdefs.h which is
not really needed.

Similarly, remove the __STDC_DEC_FP__ definition.  It should be
provided by libdfp or the toolchain, and is otherwise unused
by libdfp. The useful bits in gcc's float.h are controlled by
the presence of __STDC_WANT_DEC_FP__.

Signed-off-by: Paul E. Murphy <murp@ibm.com>