libbid: use bfp rounding mode when converting to bfp
The current behavior uses the decimal rounding mode to
round in the destination format.  This is wrong, use the
target rounding mode.  This kind of works since bfp maps 1-1
to dfp (not so in the other direction).

Inject a function in bid to stash the local decimal rounding mode
while converting to bfp.

TODO: fix fenv tests for non-existent rounding modes in BID.