expd64: correctly check for range
The original check converted to integer and checked the result.

By standards, this value is undefined if the value exceeds the
range of the integer type.  powerpc conveniently converts such
values to INT_MAX or INT_MIN.  libbid is less predictable.

Instead, just check fabs(val) is within the domain of exp
for _Decimal64.

Likewise, add missing PLT redirects for ge*2 and le*2.
The __bid_gedd2 redirect was missing and causing a failure in
the local PLT check.