libdecnumber: build and install static libdecnumber
This takes a simple approach and builds a static library
suitable for end-users.  We recycle the libdecnumber
autotooling from GCC to generate the static library in
isolation.

libdecnumber is useful for arbitrary precision decimal math,
likewise the API is fairly simplistic provided one is careful
with the variable length decNumber objects and the context
objects.

Note, libdecnumber is licensed under GPL V3 + Runtime Exception
as is the custom for software used in exported by GCC.  This
stands in constrast to other libdfp bits which are LGPLv2.1.

This gives the ability to use the three headers:

  decNumber.h decContext.h decConvert.h

decConvert.h is a header unique to libdfp to expose easy
conversions between decNumber and _DecimalN types already
present (but awkwardly declared) in libdecnumber.

to best utilize both decNumber and _DecimalN features as needed
without needing to jump through any weird hoops.

Signed-off-by: Paul E. Murphy <murphyp@linux.vnet.ibm.com>