test-bfp-conversions: refactorwrite support for _Float128
This is unwieldy and there is much crossing of the type, format,
and gcc mode streams.

Rewrite (really a hidden refactor) this to test based on GCC modes,
and enable some testing of the _Float128 support for those targets
which support it.  Note, the existing test coverage is very limited.

We only extend those tests for the float128 format to KFmode.  A
future patch may further rewrite this to better automate conversion
checking.

Likewise, workaround GCC 4.8 limitations (aka RHEL7) to continue
testing support.  It is a bit noisier about warnings for the
unfilled structures, sorry.