Add a simple build testing script
While it is impractical to run the tests on supported variants,
we can at least try to compile everything and run tests which
are not target dependent.  This is not meant to be perfect.

This relies on the build-many-glibcs.py found in glibc to construct
toolchains for all supported platforms.  For this we test the
following configurations:

s390{,x}
   - z14
   - soft-dfp
powerpc{,64}
   - power8
   - soft-dfp
x86-64
i386
powerpc64le
   - power9

These represent the actively supported variants, and compile all
tests (and run abi/plt checks too).  This leaves generating the
required toolchains from the glibc script up to the developer.

./build_all.sh ${glibc_build_many_scratch} ${libdfp_srcdir}

Signed-off-by: Paul E. Murphy <murphyp@linux.vnet.ibm.com>