Do not disambiguate configury top_srcdir
This very quietly breaks reproducible builds on debian as the
the usage of assert in printf_dfp.c will use the fully prefixed
path when expanding __FILE__ in gcc.

The original issue worked around was the broken install-doc target
which failed to invoke makefile rule target to generate ChangeLog.md
when building in subdirectory.

Instead, revert the top_srcdir changes and generate the installed
ChangeLog.md on demand if ChangeLog.md is not present in the
source tree.

This fixes issue #147

Signed-off-by: Paul E. Murphy <murphyp@linux.vnet.ibm.com>