powerpc: use __builtin_fabsdN for fabsdN
The inline asm for fabsd128 works only by strict assumption that it
is not inlined and the src and dst registers selected are identical.
When lto is used it will fail to move the lower half of the
decimal128 value causing strange failures if the value is not
computed in place.

Instead use the builtin, it's been available for quite some time
and can generate the extra move as needed.  Similarly, use the
builtins for other types as the generated code is equivalent.

This fixes the non-trivial failures discovered when turning
on lto from Issue #150.

Signed-off-by: Paul E. Murphy <murphyp@linux.vnet.ibm.com>