Add TS18661-2 strfromdN
This is a creative wrapper around the libdfp printf hook.

Fortunately, the behavior is a restricted subset of the
printf family of IO, thus keeping life simple.  The only
gotcha is that using memory files requires dynamic allocation
which could fail.

test-printf is enhanced to test the strfromd* variants under
nominal circumstances (i.e valid format and unlimited buffer).

test-strfrom is added to capture some more exotic usage of
strfrom.  That is, length computations and bogus format
strings.

Likewise, fix bug in scaffold.c which uses the wrong array
size to buffer. CHAR_MAX is both misleading, and too small
on systems where it is signed for the existing test cases.

Signed-off-by: Paul E. Murphy <murphyp@linux.vnet.ibm.com>