implement proper semantics for operator>>
This implementation mainly copies the glibc functions for normal binary
floating-point types. See below[^glibc] for a list of those functions

Note that the tests for determining whether operator>> properly parses
"inf" and "INF" have been removed

The semantics of operator>> do not call for these to be parsed
correctly. A thorough read of [do_get][1] reveals as much. Apache's
libstdcxx also [worked through this issue][2]

Tests for ensuring that operator>> consumes the expected number of
chars from the stream have been added. This did uncover some issues
with the current implementation of operator<< that will need to be
addressed

[1]: https://en.cppreference.com/w/cpp/locale/num_get/get
[2]: https://issues.apache.org/jira/browse/STDCXX-239

[^glibc]: Here are the functions and files referenced

	std::basic_istream<>::_M_extract
	std::num_get<>::do_get
	std::num_get<>::_M_extract_float
	std::__convert_to_v<>

	commit	8fbe1b18ed090a026d08f0b0e5f8cd56d652125d
	file	libstdc++-v3/include/bits/istream.tcc

	commit  8fbe1b18ed090a026d08f0b0e5f8cd56d652125d
	file    libstdc++-v3/include/bits/locale_facets.tcc

	commit  8fbe1b18ed090a026d08f0b0e5f8cd56d652125d
	file    libstdc++-v3/config/locale/generic/c_locale.cc

Signed-off-by: Dacoda Strack <dacoda.strack@gmail.com>