Enable support for AArch64 (using BID format)
I have recently added DFP support for AArch64 in GCC, and I used
libdfp as testsuite.  This required small changes:

* update of config/dfp.m4 to recognize aarch64, and regeneration of
  libdecnumber/configure

* introduction of #ifdef __LIBGCC_HAS_XF_MODE__ in libbid/*xf*.c like
  what is now in GCC gcc/libgcc/config/libbid.  This is necessary
  because AArch64 does not support XFmode.

* in libbid/bid_gcc_intrinsics.h, define TFtype to long double on
  AArch64, which does not support __float128.  This is not needed in
  GCC's libbid because BID_HAS_TF_MODE is not defined in that case.

* creation of sysdeps/aarch64/

I am not sure if the discrepancy between this libbid and GCC's copy
should be fixed, nor how.

To make the build succeed, I also had to update config.guess and
config.sub to more recent versions, but I do not include that change
in this patch because it is not strictly related to AArch64
enablement.

'make check' reports 172 PASS.

Signed-off-by: Christophe Lyon <christophe.lyon@arm.com>