ieee754r: fix pow for large exponents
Use higher precision to compute x^int(y) as this becomes important
if x^int(y) is subnormal, or large enough to compound rounding
error when computing large powers.

Likewise, rework intpow to avoid overflowing if int(y) is near
the limits of INT64_MAX.

Likewise, rework intpow to avoid undefined behavior if the
integer conversion overflows.

Likewise, assume the implementation can only be used for _Decimal64
and clean up macro indirection.

The accuracy for very large exponents is likely less than ideal,
but hopefully not entirely wrong.  I have not tested the current
implementation for accuracy, but I suspect it is lacking.