ieee754r: rewrite nexttoward
This is the same function as nextafter, but the direction is taken as
a _Decimal128.  Apply the same rewrite as nextafter, but promote the
first argument to _Decimal128 to determine direction.

Likewise, update the test suite to test this. This is apparently
the first generated test which uses different decimal type arguments,
so the python script needed some fixes.

fixes #196