ieee754r: rewrite fmod
Previously, fmod used libdecnumber's decNumberRemainder.  It turns out
decNumberRemainder gives up if the values are sufficiently far apart
and returns a NaN.  That's a serious bug/limitation.

This is derived from Nelson H.F. Beebe's book The Mathematical Function
Computation Handbook.  Unfortunately for a base > 2, an extra digit is
needed in the accumulator to ensure exact subtraction.  This can be
observed with some of the chosen test cases.

The is fairly straightfoward for _Decimal32 and _Decimal64. Just use
_Decimal128. However, _Decimal128 requires pair-precision arithmetic
when an extra-digit is required.  This implementation uses the those
functions as described by Beebe too.

Finally, for verification gen-auto-dfp-tests implements a version of
the same function using decNumber. The decNumberRemainder function
gives up if the values are too far apart.  Similarly, mpfr cannot
be used without unrealistically high precision as the radix is very
important in determining the result.